Refactor functions to modular js
