+++ src/index.html
@@ -11,20 +11,20 @@
     <nav class="navbar navbar-expand-lg navbar-light bg-light">
         <div class="container-fluid">
             <div class="collapse navbar-collapse" id="navbarSupportedContent">
-                 <a class="navbar-brand mt-2 mt-lg-0" href="#">
-                     <img
+                <a class="navbar-brand mt-2 mt-lg-0" href="#">
+                    <img
                         src="https://download.logo.wine/logo/Epic_Games/Epic_Games-Logo.wine.png"
                         height="60"
                         alt="MDB Logo"
                         loading="lazy"
           />
-                 </a>
+                </a>
           <ul class="navbar-nav me-auto mb-2 mb-lg-0">
             <li class="nav-item">
-             <a class="nav-link" href="#">Free Games (<span class="itemNav"></span>)</a>
-           </li>
-         </ul>
-       </div>
+            <a class="nav-link" href="#">Free Games (<span class="itemNav"></span>)</a>
+          </li>
+        </ul>
+      </div>
         <div class="dropdown">
           <a
             class="dropdown-toggle d-flex align-items-center hidden-arrow"
